CI: make docker storage driver implicit
For dind (docker-in-docker) jobs, we want to inherit the storage driver
configuration from the runner itself. This is necessary because storage
backends may be different between docker daemons running on separate
runner hosts.

In the runner config (/etc/gitlab-runner/config.toml) we may set this
environment variable with the "evironment" key.

eg. environment = ["DOCKER_DRIVER=zfs"]